[DEV-32] 커스텀 에러코드 적용 (#281)
* refactor: custom errorCode 정의

* feat: ExceptionResponse 스펙 수정

* refactor: ErrorCode 추가 및 적용

* refactor: Auth - 로그인 요청 custom errorCode 적용

* refactor: 수강과목 등록 - pdf parsing text custom errorCode 적용

* feat: StudentCategory - 학생 이수구분 카테고리에 졸업 카테고리가 포함 여부 검증 로직 추가

* feat: 유저아이디 찾기 커스텀 에러코드 적용

* feat: 유저검증 커스텀 에러코드 적용

* feat: 비밀번호 재설정 커스텀 에러코드 적용

* feat: 아이디 중복 체크 커스텀 에러코드 적용

* feat: 학번 중복 체크정 커스텀 에러코드 적용

* feat: 회원 탈퇴 커스텀 에러코드 적용

* feat: 회원가입 요청 커스텀 에러코드 적용

* refactor: GlobalExceptionHandler response 스펙 수정

* refactor: 불필요 import 제거

+++ src/main/java/com/plzgraduate/myongjigraduatebe/user/application/service/withdraw/WithDrawUserService.java
@@ -4,6 +4,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.plzgraduate.myongjigraduatebe.completedcredit.application.port.DeleteCompletedCreditPort;
+import com.plzgraduate.myongjigraduatebe.core.exception.ErrorCode;
 import com.plzgraduate.myongjigraduatebe.core.meta.UseCase;
 import com.plzgraduate.myongjigraduatebe.parsing.application.port.DeleteParsingTextHistoryPort;
 import com.plzgraduate.myongjigraduatebe.takenlecture.application.usecase.delete.DeleteTakenLectureUseCase;
@@ -29,7 +30,9 @@ class WithDrawUserService implements WithDrawUserUseCase {
 	@Override
 	public void withDraw(Long userId, String password) {
 		User user = findUserUseCase.findUserById(userId);
-		user.matchPassword(passwordEncoder, password);
+		if(!user.matchPassword(passwordEncoder, password)){
+			throw new IllegalArgumentException(ErrorCode.INCORRECT_PASSWORD.toString());
+		}
 		deleteTakenLectureByUserUseCase.deleteAllTakenLecturesByUser(user);
 		deleteParsingTextHistoryPort.deleteUserParsingTextHistory(user);
 		deleteCompletedCreditPort.deleteAllCompletedCredits(user);
